refactor: move contexts folder to components
