+++ src/components/molecules/multi-select/MultiSelect.controller.ts
@@ -1,7 +1,7 @@
-import { FormContext } from '@contexts';
 import { useResetFormInput } from '@hooks';
 import type { MultiSelectValue, SelectValue } from '@types';
 import { useCallback, useContext, useRef, useState } from 'react';
+import { FormContext } from '../../contexts';
 
 import {
   getMultiSelectFormValues,
