+++ src/components/molecules/switch/__tests__/Switch.controller.test.tsx
@@ -1,8 +1,8 @@
-import { FormContext } from '@contexts';
 import { act, renderHook } from '@testing-library/react';
 import type { FormContextProps } from '@types';
 import { type KeyboardEvent, useEffect } from 'react';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
+import { FormContext } from '../../../contexts';
 import { useSwitchController } from '../Switch.controller';
 
 // Mock the dependencies
