+++ src/hooks/use-theme/use-chop-logic-theme.ts
@@ -1,4 +1,4 @@
-import { ThemeContext } from '@contexts';
+import { ThemeContext } from '@components/contexts';
 import { useContext } from 'react';
 
 export const useTheme = () => useContext(ThemeContext);
