+++ src/index.ts
@@ -1,8 +1,8 @@
 export * from './components/atoms';
+export * from './components/contexts';
 export * from './components/hocs';
 export * from './components/molecules';
 export * from './components/organisms';
-export * from './contexts';
 export * from './enums';
 export * from './hooks';
 export * from './types';
